requirements.txt
-e svn+http://code.djangoproject.com/svn/django/trunk@17672#egg=Django-1.4c1-py2.7-dev_r17672
MySQL-python==1.2.3
distribute==0.6.24
django-extensions==0.8
virtualenvwrapper==2.11.1
wsgiref==0.1.2
django-tastypie==0.9.11
